server/requirements.txt
# Connectivity, Encryption and Statistics Gathering
paramiko==2.4.0
psutil==5.4.2
pycrypto==2.6.1

# Data and workers.
SQLAlchemy==1.1.15
alembic==0.9.6
mysqlclient==1.3.12
celery==4.1


# Testing and static analysis.
pytest==3.3.1
pytest-cov==2.5.1
